requirements.txt
torchvision
tensorboardX
